refactor CustomCollectionConditionTest to use new methods `check` and `fail`
